Moved rgettatr fn to mixins
