+++ commercialoperator/components/segregation/mixins.py
@@ -1,3 +1,5 @@
+import functools
+from commercialoperator.components.segregation.decorators import basic_exception_handler
 from commercialoperator.components.segregation.utils import (
     retrieve_email_user,
     retrieve_group_members,
@@ -24,3 +26,16 @@ class MembersPropertiesMixin:
         all_members.extend(retrieve_group_members(group_object=self))
         emailuser = [retrieve_email_user(m) for m in all_members]
         return [u.email for u in emailuser if u]
+
+
+class RecursiveGetAttributeMixin:
+
+    @basic_exception_handler
+    def rgetattr(self, obj, attr, *args):
+        def _getattr(obj, attr):
+            if isinstance(obj, dict):
+                return obj.get(attr, None)
+            else:
+                return getattr(obj, attr, *args) if hasattr(obj, attr) else None
+
+        return functools.reduce(_getattr, [obj] + attr.split("."))
